fixed product controller, services and rout

+++ services/product.service.js
@@ -13,12 +13,10 @@ class ProductService {
         const filter = {};
 
         if (category) {
-
             filter.category = category;
         }
 
-        return await Product.find(filter)
-            .sort({ createdAt: -1 });
+        return await Product.find(filter).sort({ createdAt: -1 });
     }
 
     async getById(id) {
@@ -26,14 +24,23 @@ class ProductService {
     }
 
     async update(id, data) {
-        return await Product.findByIdAndUpdate(
-            id,
-            data,
-            { new: true }
-        );
+        const existing = await Product.findById(id);
+        if (!existing) {
+            const error = new Error('PRODUCT_NOT_FOUND');
+            error.code = 'PRODUCT_NOT_FOUND';
+            throw error;
+        }
+        return await Product.findByIdAndUpdate(id,data,{ new: true });
     }
 
     async delete(id) {
+        const existing = await Product.findById(id);
+        if (!existing) {
+            const error = new Error('PRODUCT_NOT_FOUND');
+            error.code = 'PRODUCT_NOT_FOUND';
+            throw error;
+        }
+
         return await Product.findByIdAndDelete(id);
     }
 
@@ -48,18 +55,24 @@ class ProductService {
         const user = await User.findById(userId);
 
         if(!user){
-            throw new Error('User not found');
+            const error = new Error('USER_NOT_FOUND');
+            error.code = 'USER_NOT_FOUND';
+            throw error;
         }
 
         const product = await Product.findById(productId);
 
         if (!product){
-            throw new Error('Product not found');
+            const error = new Error('PRODUCT_NOT_FOUND');
+            error.code = 'PRODUCT_NOT_FOUND';
+            throw error;
         }
 
         if(user.money < product.price){
 
-            throw new Error('Not enough money');
+            const error = new Error('INSUFFICIENT_FUNDS');
+            error.code = 'INSUFFICIENT_FUNDS';
+            throw error;
         }
 
         user.money -= product.price;
@@ -78,7 +91,7 @@ class ProductService {
 
         await user.save();
 
-        return user, product;
+        return {user, product};
     }
 
     async getUserInventory(userId) {
